Move renderer test into its own suite
